Modify the control of the notification to be able send it to driver and admin
